Rename `setup.py` to `conftest.py` and move test fixtures into it.
